requirements.txt
psycopg2-binary
django-haystack
libsass
elasticsearch
django-bootstrap4
django-allauth
python-decouple==3.1
django_filter==2.2.0
django-widget-tweaks==1.4.2
django-debug-toolbar